pyproject.toml
[tool.poetry]
name = "coinspro-py"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "coinspro_py"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
websockets = "^11.0.3"
websocket-client = "^1.6.2"
python-dotenv = "^1.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
